Rephrase batch table parsing as a function
