+++ Specs/Scene/parseBatchTableSpec.js
@@ -1,0 +1,56 @@
+/*
+var x = {
+ schema: {
+   classes: {
+     _batchTable: {
+       properties: {}
+     }
+   },
+   featureTables: {
+     _batchTable: {
+       class: "_batchTable",
+       count: 10
+       properties: {}
+     }
+   }
+ }
+};
+*/
+
+describe("Scene/parseBatchTable", function () {
+  it("throws without count", function () {
+    fail();
+  });
+
+  it("throws without batchTable", function () {
+    fail();
+  });
+
+  it("throws without binary body", function () {
+    fail();
+  });
+
+  it("warns for deprecated HIERARCHY extension", function () {
+    fail();
+  });
+
+  it("parses batch table with no properties", function () {
+    fail();
+  });
+
+  it("parses batch table with binary properties", function () {
+    fail();
+  });
+
+  it("parses batch table with JSON properties", function () {
+    fail();
+  });
+
+  it("parses batch table with hierarchy", function () {
+    fail();
+  });
+
+  it("stores extras and extensions in the transcoded FeatureTable", function () {
+    fail();
+  });
+});
